upates

+++ database/migrations/2017_04_24_160036_create_settings_table.php
@@ -61,7 +61,7 @@ class CreateSettingsTable extends Migration
             $table->text('section_three_en')->nullable();
             $table->string('video')->nullable();
             $table->tinyInteger('service_versions_limit')->nullable();
-            $table->tinyInteger('initial_points')->nullable()->default(0);
+            $table->integer('initial_points')->nullable()->default(0);
             $table->boolean('annual_subscription')->default(0);
             $table->boolean('maintenance_mode')->default(0);
             $table->boolean('auto_enrollment')->default(0);
